requirements.txt
pandas
mysql-connector-python
gspread
oauth2client
python-dotenv
SQLAlchemy
pymysql
